remove unhandledrejection listener and provide setError in a context
Listening to all unhandledrejection events is not ideal for setting the
error screen, as these events can also come from non-app code, which
leads to the error screen displaying for non-issues from third-party
code. This updates the ErrorBoundary to remove the listener
completely, and provide setError as the preferred method for handling
async errors where they occur.